# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trak fitness coaching platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Invite Links ──────────────────────────────────────────────────────────
  /clients/{clientId}/invite:
    post:
      operationId: generateInviteLink
      tags: [invites]
      summary: Generate or refresh an invite link for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLink"
  /invite/{token}:
    get:
      operationId: getInvite
      tags: [invites]
      summary: Get invite info by token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteInfo"
        "404":
          description: Invalid or expired token
  /invite/{token}/accept:
    post:
      operationId: acceptInvite
      tags: [invites]
      summary: Accept an invite as the currently signed-in Clerk user
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteAcceptResult"
        "401":
          description: Not signed in
        "403":
          description: Email mismatch or already claimed by another account
        "404":
          description: Invalid or expired token
  /clients/me:
    get:
      operationId: getMyClient
      tags: [clients]
      summary: Get the client record for the currently signed-in Clerk user
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Not signed in
        "404":
          description: No client account found for this user

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
  /clients/{clientId}/coach-notes:
    get:
      operationId: listCoachNotes
      tags: [coach]
      summary: List private coach notes for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coach notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachNote"
    post:
      operationId: createCoachNote
      tags: [coach]
      summary: Create a private coach note
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachNoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachNote"
  /clients/{clientId}/coach-notes/{noteId}:
    patch:
      operationId: updateCoachNote
      tags: [coach]
      summary: Update a coach note
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachNoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachNote"
    delete:
      operationId: deleteCoachNote
      tags: [coach]
      summary: Delete a coach note
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Call Logs ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        inviteToken:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        goal:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    InviteLink:
      type: object
      required: [token, url]
      properties:
        token:
          type: string
        url:
          type: string
    InviteInfo:
      type: object
      required: [clientId, clientName, clientEmail]
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
    InviteAcceptResult:
      type: object
      required: [clientId, clientName, alreadyJoined]
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        alreadyJoined:
          type: boolean

    # ── Dashboard ───────────────────────────────────────────────────────────
...
    CoachNote:
      type: object
      required: [id, clientId, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CoachNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...