# lib/api-zod/src/generated/types/inviteAcceptResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trak fitness coaching platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InviteAcceptResult {
  clientId: number;
  clientName: string;
  alreadyJoined: boolean;
}